+++ src/Skeletest/Internal/TestRunner.hs
@@ -1,4 +1,6 @@
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE LambdaCase #-}
+{-# LANGUAGE OverloadedRecordDot #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE NoFieldSelectors #-}
@@ -9,6 +11,7 @@ module Skeletest.Internal.TestRunner (
 
   -- * TestResult
   TestResult (..),
+  TestResultStatus (..),
   TestResultMessage (..),
   testResultPass,
   testResultFromAssertionFail,
@@ -26,6 +29,7 @@ import Data.Text (Text)
 import Data.Text qualified as Text
 import Data.Typeable (typeOf)
 import GHC.IO.Exception qualified as GHC
+import GHC.Records (HasField (..))
 import GHC.Stack (CallStack)
 import GHC.Stack qualified as GHC
 import Skeletest.Internal.Error (SkeletestError)
@@ -59,14 +63,36 @@ class (MonadIO m) => Testable m where
 
 {----- TestResult -----}
 
--- TODO: Remove 'testResult' prefix
 -- TODO: Change success -> PASS/FAIL/SKIP
 -- https://github.com/brandonchinn178/skeletest/issues/89
 data TestResult = TestResult
-  { testResultSuccess :: Bool
-  , testResultLabel :: Text
-  , testResultMessage :: TestResultMessage
+  { status :: TestResultStatus
+  , label :: Text
+  , message :: TestResultMessage
   }
+
+data TestResultStatus
+  = TestPassed
+  | TestFailed
+  | TestSkipped
+  | TestStatus
+      { name_ :: Text
+      , success_ :: Bool
+      }
+  deriving (Eq, Ord)
+
+instance HasField "name" TestResultStatus Text where
+  getField = \case
+    TestPassed -> "passed"
+    TestFailed -> "failed"
+    TestSkipped -> "skipped"
+    TestStatus{name_} -> name_
+instance HasField "success" TestResultStatus Bool where
+  getField = \case
+    TestPassed -> True
+    TestFailed -> False
+    TestSkipped -> True
+    TestStatus{success_} -> success_
 
 data TestResultMessage
   = TestResultMessageNone
@@ -76,9 +102,9 @@ data TestResultMessage
 testResultPass :: TestResult
 testResultPass =
   TestResult
-    { testResultSuccess = True
-    , testResultLabel = Color.green "OK"
-    , testResultMessage = TestResultMessageNone
+    { status = TestPassed
+    , label = Color.green "OK"
+    , message = TestResultMessageNone
     }
 
 testResultFromAssertionFail :: AssertionFail -> IO TestResult
@@ -86,9 +112,9 @@ testResultFromAssertionFail e = do
   msg <- renderAssertionFail e
   pure
     TestResult
-      { testResultSuccess = False
-      , testResultLabel = Color.red "FAIL"
-      , testResultMessage = TestResultMessageBox [BoxText msg]
+      { status = TestFailed
+      , label = Color.red "FAIL"
+      , message = TestResultMessageBox [BoxText msg]
       }
 
 testResultFromError :: SomeException -> IO TestResult
@@ -99,9 +125,9 @@ testResultFromErrorWith f e = do
   msg <- f <$> renderMsg
   pure
     TestResult
-      { testResultSuccess = False
-      , testResultLabel = Color.red "ERROR"
-      , testResultMessage = TestResultMessageBox [BoxText msg]
+      { status = TestFailed
+      , label = Color.red "ERROR"
+      , message = TestResultMessageBox [BoxText msg]
       }
  where
   renderMsg